Add utils for login as admin (#19223)
* add utils for login as admin

* update as per comments

* fix java checkstyle issue

---------
